admin register faculty (#40)
* Add /api/login endpoint

Login takes a username and password
Passes to existing `User->login` and verifies
that users exists in DB.

Returns a JWT with uid/role

* Add view for inviting faculty to register

Add view for inviting faculty to register
Backend not implemented. Asks for faculty,
email, and department which will be prefilled
when the faculty visits the registration form.

* Allow admin to send invitation email

Adds UI and backend for faculty to enter
a new faculty's name/email/department
and the system sends them an invitation email
with a link to complete their registration

* Break header and message box into their components

Just a little component clean up. App.vue was getting too complex

* Parse out base64 encoded to prefill registration form

When a new user is sent a registration link that link
contains a base64 encoded string with their name, email,
and department so when they visit the registration page
that info is prefilled

* Pull admin email from db before sending email
* Add startup fund to faculty invitation form